[Refactor/#73] API Document 업데이트
