+++ src/main/java/life/inha/icemarket/config/swagger/LoginAPIDocument.java
@@ -1,10 +1,8 @@
-package life.inha.icemarket.controller;
+package life.inha.icemarket.config.swagger;
 
 import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import life.inha.icemarket.config.swagger.ApiDocumentResponse;
-import org.springframework.http.MediaType;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
@@ -15,7 +13,7 @@ import org.springframework.web.bind.annotation.RequestParam;
  */
 @Tag(name="Login", description = "로그인 API")
 @Controller
-public class AuthApi {
+public class LoginAPIDocument {
     @Operation(description = "로그인 페이지 - email과 비밀번호를 param으로 받아오면 jwt 토큰값을 리턴합니다.")
     @ApiDocumentResponse
     @RequestMapping(value = "/login ", method = RequestMethod.POST)
