+++ src/main/java/life/inha/icemarket/controller/FindPasswordController.java
@@ -71,7 +71,7 @@ public class FindPasswordController {
 
 
     // 비밀번호 찾기 get 요청 컨트롤러
-    @Operation(description = "비밀번호 찾기 페이지")
+    @Operation(description = "비밀번호 찾기 페이지 - find_pw.html에 FindPasswordForm을 담아서 보냅니다.")
     @ApiDocumentResponse
     @GetMapping("/findpw")
     public String findpw(Model model){
@@ -81,7 +81,7 @@ public class FindPasswordController {
 
 
     // 비밀번호 찾기 post 요청 컨트롤러 시작 //
-    @Operation(description = "비밀번호 찾기")
+    @Operation(description = "POST 비밀번호 찾기 - 채워진 FindPasswordForm을 받아 비밀번호를 바꿀 사용자를 확인한 뒤 해당 사용자의 이메일을 담은 ResetPasswordForm을 resetpw.html에 보냅니다.")
     @ApiDocumentResponse
    // @ResponseBody
     @RequestMapping(
@@ -112,7 +112,7 @@ public class FindPasswordController {
     // 비밀번호 찾기 post 요청 컨트롤러 끝 //
 
     // 비밀번호 초기화 post 요청 컨트롤러 시작 //
-    @Operation(description = "비밀번호 초기화")
+    @Operation(description = "비밀번호 초기화 - 채워진 ResetPasswordForm을 받아 비밀번호를 변경합니다.")
     @ApiDocumentResponse
     @ResponseBody
     @RequestMapping(
